+++ packages/provider-queries/src/get-tokenised-article-url.js
@@ -1,0 +1,10 @@
+import { addTypenameToDocument } from "apollo-utilities";
+import gql from "graphql-tag";
+
+export default addTypenameToDocument(gql`
+  query GetTokenisedArticleUrl($id: ID!) {
+    article(id: $id) {
+      tokenisedUrl
+    }
+  }
+`);