+++ packages/ssr/src/component/author-profile.js
@@ -7,6 +7,7 @@ const { AuthorProfileProvider } = require("@times-components/provider/rnw");
 const {
   ContextProviderWithDefaults
 } = require("@times-components/context/rnw");
+const { StickyProvider } = require("@times-components/sticky/rnw");
 const AuthorProfile = require("@times-components/author-profile/rnw").default;
 
 module.exports = (client, analyticsStream, data, helmetContext) => {
@@ -21,37 +22,41 @@ module.exports = (client, analyticsStream, data, helmetContext) => {
   } = data;
 
   return React.createElement(
-    HelmetProvider,
-    { context: helmetContext },
+    StickyProvider,
+    {},
     React.createElement(
-      ApolloProvider,
-      { client },
+      HelmetProvider,
+      { context: helmetContext },
       React.createElement(
-        AuthorProfileProvider,
-        {
-          debounceTimeMs,
-          page,
-          pageSize,
-          slug: authorSlug
-        },
-        ({ author, isLoading, error, refetch }) =>
-          React.createElement(
-            ContextProviderWithDefaults,
-            { value: { makeArticleUrl, makeTopicUrl } },
-            React.createElement(AuthorProfile, {
-              adConfig: mapProfileToAdConfig(),
-              analyticsStream,
-              author,
-              error,
-              isLoading,
-              onArticlePress: () => {},
-              onTwitterLinkPress: () => {},
-              page,
-              pageSize,
-              refetch,
-              slug: authorSlug
-            })
-          )
+        ApolloProvider,
+        { client },
+        React.createElement(
+          AuthorProfileProvider,
+          {
+            debounceTimeMs,
+            page,
+            pageSize,
+            slug: authorSlug
+          },
+          ({ author, isLoading, error, refetch }) =>
+            React.createElement(
+              ContextProviderWithDefaults,
+              { value: { makeArticleUrl, makeTopicUrl } },
+              React.createElement(AuthorProfile, {
+                adConfig: mapProfileToAdConfig(),
+                analyticsStream,
+                author,
+                error,
+                isLoading,
+                onArticlePress: () => {},
+                onTwitterLinkPress: () => {},
+                page,
+                pageSize,
+                refetch,
+                slug: authorSlug
+              })
+            )
+        )
       )
     )
   );