+++ packages/article-byline/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.75](https://github.com/newsuk/times-components/compare/@times-components/article-byline@3.1.74...@times-components/article-byline@3.1.75) (2019-05-31)
+
+**Note:** Version bump only for package @times-components/article-byline
+
+
+
+
+
 ## [3.1.74](https://github.com/newsuk/times-components/compare/@times-components/article-byline@3.1.73...@times-components/article-byline@3.1.74) (2019-05-28)
 
 **Note:** Version bump only for package @times-components/article-byline