+++ packages/article-label/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.82](https://github.com/newsuk/times-components/compare/@times-components/article-label@2.3.81...@times-components/article-label@2.3.82) (2019-05-31)
+
+**Note:** Version bump only for package @times-components/article-label
+
+
+
+
+
 ## [2.3.81](https://github.com/newsuk/times-components/compare/@times-components/article-label@2.3.80...@times-components/article-label@2.3.81) (2019-05-23)
 
 **Note:** Version bump only for package @times-components/article-label