+++ packages/message-bar/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.9](https://github.com/newsuk/times-components/compare/@times-components/message-bar@0.2.8...@times-components/message-bar@0.2.9) (2019-05-31)
+
+
+### Bug Fixes
+
+* message bar consumed and styled in web ([#2003](https://github.com/newsuk/times-components/issues/2003)) ([7c5c65e](https://github.com/newsuk/times-components/commit/7c5c65e))
+
+
+
+
+
 ## [0.2.8](https://github.com/newsuk/times-components/compare/@times-components/message-bar@0.2.7...@times-components/message-bar@0.2.8) (2019-05-23)
 
 **Note:** Version bump only for package @times-components/message-bar