+++ packages/provider-test-tools/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.17.18](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@1.17.17...@times-components/provider-test-tools@1.17.18) (2019-05-31)
+
+**Note:** Version bump only for package @times-components/provider-test-tools
+
+
+
+
+
 ## [1.17.17](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@1.17.16...@times-components/provider-test-tools@1.17.17) (2019-05-24)
 
 **Note:** Version bump only for package @times-components/provider-test-tools